--- Solan/frontend/src/styling/toolbar.css
@@ -1,60 +1,0 @@
-.toolbar {
-    position:fixed;
-    width: 100%;
-    height: 6rem;
-    background: #000000;
-}
-
-
-.toolbar_navigation {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    padding:  0 1rem;
-
-}
-
-.button {
-
-    height: 1.8rem;
-    background-color: rgb(255, 207, 171);
-    border: none;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    border-radius: 30px;
-    font-size: 16px;
-    cursor: pointer;
-}
-
-.toolbar_logo a {
-    color: white;
-    text-decoration: none;
-    font-size: 1.5rem;
-}
-
-.toolbar_navigation-items a {
-    font-size: 18px;
-    color: white;
-    text-decoration: none;
-}
-
-
-.toolbar_navigation-items ul {
-    align-items: center;
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    display: flex;
-}
-
-.spacer {
-    flex: 1;
-}
-
-.toolbar_navigation-items li {
-    padding: 0 0.5rem;
-}
-
-
